+++ src/Umbraco.Tests.Integration/RuntimeTests.cs
@@ -15,9 +15,9 @@ using Umbraco.Tests.Common;
 using Umbraco.Tests.Integration.Extensions;
 using Umbraco.Tests.Integration.Implementations;
 using Umbraco.Tests.Integration.Testing;
-using Umbraco.Web.BackOffice.AspNetCore;
 using Umbraco.Web.Common.AspNetCore;
 using Umbraco.Extensions;
+using Umbraco.Web.BackOffice.Extensions;
 
 namespace Umbraco.Tests.Integration
 {
