+++ src/Umbraco.Web.Common/Extensions/ApplicationBuilderExtensions.cs
@@ -1,8 +1,12 @@
 using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
+using Serilog.Context;
+using Smidge;
 using StackExchange.Profiling;
 using Umbraco.Core;
+using Umbraco.Core.Hosting;
+using Umbraco.Infrastructure.Logging.Serilog.Enrichers;
 using Umbraco.Web.Common.Middleware;
 
 namespace Umbraco.Extensions
@@ -22,6 +26,59 @@ namespace Umbraco.Extensions
         }
 
         /// <summary>
+        /// Enables the Umbraco installer
+        /// </summary>
+        /// <param name="app"></param>
+        /// <returns></returns>
+        public static IApplicationBuilder UseUmbracoInstaller(this IApplicationBuilder app)
+        {
+            app.UseEndpoints(endpoints =>
+            {
+                // TODO: Fix this routing with an area
+                endpoints.MapControllerRoute("Install", "/install/{controller}/{Action}", defaults: new { Area = "Install" });
+
+                //TODO register routing correct: Name must be like this
+                endpoints.MapControllerRoute("umbraco-api-UmbracoInstall-InstallApi", "/install/api/{Action}", defaults: new { Area = "Install", Controller = "InstallApi" });
+            });
+
+            return app;
+        }
+
+        /// <summary>
+        /// Start Umbraco
+        /// </summary>
+        /// <param name="app"></param>
+        /// <returns></returns>
+        public static IApplicationBuilder UseUmbracoCore(this IApplicationBuilder app)
+        {
+            if (app == null) throw new ArgumentNullException(nameof(app));
+
+            if (app.UmbracoCanBoot())
+            {
+                var runtime = app.ApplicationServices.GetRequiredService<IRuntime>();
+
+                // Register a listener for application shutdown in order to terminate the runtime
+                var hostLifetime = app.ApplicationServices.GetRequiredService<IApplicationShutdownRegistry>();
+                var runtimeShutdown = new CoreRuntimeShutdown(runtime, hostLifetime);
+                hostLifetime.RegisterObject(runtimeShutdown);
+
+                // Register our global threadabort enricher for logging
+                var threadAbortEnricher = app.ApplicationServices.GetRequiredService<ThreadAbortExceptionEnricher>();
+                LogContext.Push(threadAbortEnricher); // NOTE: We are not in a using clause because we are not removing it, it is on the global context
+
+                // Start the runtime!
+                runtime.Start();
+            }
+            else
+            {
+                // TODO: Register simple middleware to show the error like we used to in UmbracoModule? Or maybe that's part of a UseUmbracoWebsite/backoffice type thing .. probably :)
+
+            }
+
+            return app;
+        }
+
+        /// <summary>
         /// Enables middlewares required to run Umbraco
         /// </summary>
         /// <param name="app"></param>
@@ -36,6 +93,60 @@ namespace Umbraco.Extensions
             app.UseMiddleware<UmbracoRequestMiddleware>();
             app.UseMiddleware<MiniProfilerMiddleware>();
             return app;
+        }
+
+        public static IApplicationBuilder UseUmbracoRequestLogging(this IApplicationBuilder app)
+        {
+            if (app == null) throw new ArgumentNullException(nameof(app));
+
+            if (!app.UmbracoCanBoot()) return app;
+
+            app.UseMiddleware<UmbracoRequestLoggingMiddleware>();
+
+            return app;
+        }
+
+        /// <summary>
+        /// Enables runtime minification for Umbraco
+        /// </summary>
+        /// <param name="app"></param>
+        /// <returns></returns>
+        public static IApplicationBuilder UseUmbracoRuntimeMinification(this IApplicationBuilder app)
+        {
+            if (app == null) throw new ArgumentNullException(nameof(app));
+
+            if (!app.UmbracoCanBoot()) return app;
+
+            app.UseSmidge();
+
+            return app;
+        }
+
+        /// <summary>
+        /// Ensures the runtime is shutdown when the application is shutting down
+        /// </summary>
+        private class CoreRuntimeShutdown : IRegisteredObject
+        {
+            public CoreRuntimeShutdown(IRuntime runtime, IApplicationShutdownRegistry hostLifetime)
+            {
+                _runtime = runtime;
+                _hostLifetime = hostLifetime;
+            }
+
+            private bool _completed = false;
+            private readonly IRuntime _runtime;
+            private readonly IApplicationShutdownRegistry _hostLifetime;
+
+            public void Stop(bool immediate)
+            {
+                if (!_completed)
+                {
+                    _completed = true;
+                    _runtime.Terminate();
+                    _hostLifetime.UnregisterObject(this);
+                }
+
+            }
         }
     }
