+++ src/Umbraco.Web.Common/Middleware/UmbracoRequestMiddleware.cs
@@ -5,46 +5,46 @@ using Microsoft.AspNetCore.Http.Extensions;
 using Umbraco.Web.Common.Lifetime;
 using Umbraco.Core;
 using Umbraco.Core.Logging;
+using System.Threading;
 
 namespace Umbraco.Web.Common.Middleware
 {
+
     /// <summary>
     /// Manages Umbraco request objects and their lifetime
     /// </summary>
-    public class UmbracoRequestMiddleware
+    public class UmbracoRequestMiddleware : IMiddleware
     {
-        private readonly RequestDelegate _next;
         private readonly ILogger _logger;
         private readonly IUmbracoRequestLifetimeManager _umbracoRequestLifetimeManager;
         private readonly IUmbracoContextFactory _umbracoContextFactory;
 
-        public UmbracoRequestMiddleware(RequestDelegate next,
+        public UmbracoRequestMiddleware(
             ILogger logger,
             IUmbracoRequestLifetimeManager umbracoRequestLifetimeManager,
             IUmbracoContextFactory umbracoContextFactory)
         {
-            _next = next;
             _logger = logger;
             _umbracoRequestLifetimeManager = umbracoRequestLifetimeManager;
             _umbracoContextFactory = umbracoContextFactory;
         }
 
-        public async Task InvokeAsync(HttpContext context)
+        public async Task InvokeAsync(HttpContext context, RequestDelegate next)
         {
             // do not process if client-side request
 
             if (new Uri(context.Request.GetEncodedUrl(), UriKind.RelativeOrAbsolute).IsClientSideRequest())
             {
-                await _next(context);
+                await next(context);
                 return;
             }
 
             var umbracoContextReference = _umbracoContextFactory.EnsureUmbracoContext();
-            
+
             try
             {
                 try
-                {                    
+                {
                     _umbracoRequestLifetimeManager.InitRequest(context);
                 }
                 catch (Exception ex)
@@ -53,7 +53,7 @@ namespace Umbraco.Web.Common.Middleware
                     _logger.Error<UmbracoRequestMiddleware>(ex);
                 }
 
-                await _next(context);
+                await next(context);
 
                 _umbracoRequestLifetimeManager.EndRequest(context);
             }
