+++ src/Umbraco.Web.Website/Extensions/UmbracoWebsiteApplicationBuilderExtensions.cs
@@ -4,10 +4,11 @@ using Microsoft.Extensions.DependencyInjection;
 using SixLabors.ImageSharp.Web.DependencyInjection;
 using Umbraco.Core;
 using Umbraco.Extensions;
+using Umbraco.Web.Common.Middleware;
 
-namespace Umbraco.Web.Website.AspNetCore
+namespace Umbraco.Web.Website.Extensions
 {
-    public static class UmbracoBackOfficeApplicationBuilderExtensions
+    public static class UmbracoWebsiteApplicationBuilderExtensions
     {
         public static IApplicationBuilder UseUmbracoWebsite(this IApplicationBuilder app)
         {
@@ -16,6 +17,7 @@ namespace Umbraco.Web.Website.AspNetCore
             if (!app.UmbracoCanBoot()) return app;
 
             // Important we handle image manipulations before the static files, otherwise the querystring is just ignored.
+            // TODO: Since we are dependent on these we need to register them but what happens when we call this multiple times since we are dependent on this for UseUmbracoBackOffice too?
             app.UseImageSharp();
             app.UseStaticFiles();
